Split providers/github.rs into focused submodules
github/mod.rs keeps the Forge impl over the gh CLI plus remote-name
parsing and error classification; github/api.rs owns the REST
operations, payload/endpoint builders, and API-to-canonical type
mapping; github/transport.rs owns the gh-api/native-HTTP dispatch,
token lookup, and the IPv4-first resolver. Pure moves, no behavior
change.

Knit-Group: kg_20260611_1e0065
Knit-Bundle: open-source-prep
